requirements.txt
alabaster==0.7.10
argh==0.26.2
Babel==2.5.3
certifi==2018.1.18
chardet==3.0.4
Django==2.0.2
django-autofixture==0.12.1
django-extensions==2.0.0
docutils==0.14
flake8==3.5.0
idna==2.6
imagesize==1.0.0
Jinja2==2.10
livereload==2.5.1
MarkupSafe==1.0
mccabe==0.6.1
packaging==17.1
pathtools==0.1.2
port-for==0.3.1
psycopg2==2.7.4
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pyparsing==2.2.0
python-decouple==3.1
pytz==2018.3
PyYAML==3.12
requests==2.18.4
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.1
sphinx-autobuild==0.7.1
sphinx-rtd-theme==0.2.4
sphinxcontrib-websupport==1.0.1
tornado==4.5.3
typing==3.6.4
urllib3==1.22
watchdog==0.8.3
